Include support for dash.conf
Include support for dash.conf,
this will allow the user to setup auto running files when dash opens
and will allow users to input commands to dash via a file.
This is useful for example if some software should be allowed to
take control over user inputs, such as a desktop environment. The file that causes user inputs is limited to the same restrictions as the user, so it still has to be root for root level calls.